refactor: dependency cycles in common match
