+++ common/match/pool.ts
@@ -2,9 +2,9 @@ import { prisma } from '../prisma';
 import type { Prisma, pupil as Pupil, student as Student } from '@prisma/client';
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-ignore The matching algorithm is optional, to allow for slim local setups
-import type { Helpee, Helper, Settings, SubjectWithGradeRestriction } from 'corona-school-matching';
+import type { Helpee, Helper, Settings } from 'corona-school-matching';
 import { createMatch } from './create';
-import { parseSubjectString, Subject } from '../util/subjectsutils';
+import { parseSubjectString } from '../util/subjectsutils';
 import { gradeAsInt } from '../util/gradestrings';
 import { assertExists } from '../util/basic';
 import { getLogger } from '../logger/logger';
@@ -13,11 +13,7 @@ import { cleanupUnconfirmed, InterestConfirmationStatus, requestInterestConfirma
 import { userSearch } from '../user/search';
 import { addPupilScreening } from '../pupil/screening';
 import assert from 'assert';
-
-export const DEFAULT_TUTORING_GRADERESTRICTIONS = {
-    MIN: 1,
-    MAX: 13,
-};
+import { formattedSubjectToSubjectWithGradeRestriction } from './util';
 
 const logger = getLogger('MatchingPool');
 
@@ -147,16 +143,6 @@ async function pupilToHelpee(pupil: Pupil): Promise<Helpee> {
         matchingPriority: pupil.matchingPriority,
         grade: gradeAsInt(pupil.grade),
         // firstMatchRequest: pupil.firstMatchRequest
-    };
-}
-
-function formattedSubjectToSubjectWithGradeRestriction(subject: Subject): SubjectWithGradeRestriction {
-    return {
-        name: subject.name,
-        gradeRestriction: {
-            min: subject.grade?.min ?? DEFAULT_TUTORING_GRADERESTRICTIONS.MIN, //due to a screening tool's bug (or how it is designed), those values may be null (which causes the algorithm to fail)
-            max: subject.grade?.max ?? DEFAULT_TUTORING_GRADERESTRICTIONS.MAX,
-        },
     };
 }
